Add matches module
`matches` module contains `rewrite_match` and related stuffs.
